daemon: move transaction `jsonrpc_` methods to an individual module
