+++ lbry/extras/daemon/daemon.py
@@ -229,6 +229,7 @@ from lbry.extras.daemon.daemon_channel import Daemon_channel
 from lbry.extras.daemon.daemon_stream import Daemon_stream
 from lbry.extras.daemon.daemon_collection import Daemon_collection
 from lbry.extras.daemon.daemon_support import Daemon_support
+from lbry.extras.daemon.daemon_transaction import Daemon_transaction
 
 HISTOGRAM_BUCKETS = (
     .005, .01, .025, .05, .075, .1, .25, .5, .75, 1.0, 2.5, 5.0, 7.5, 10.0, 15.0, 20.0, 30.0, 60.0, float('inf')
@@ -238,7 +239,8 @@ HISTOGRAM_BUCKETS = (
 class Daemon(Daemon_base, Daemon_get, Daemon_settings, Daemon_wallet,
              Daemon_account, Daemon_sync, Daemon_address, Daemon_file,
              Daemon_purchase, Daemon_claim, Daemon_channel,
-             Daemon_stream, Daemon_collection, Daemon_support):
+             Daemon_stream, Daemon_collection, Daemon_support,
+             Daemon_transaction):
     """
     LBRYnet daemon, a jsonrpc interface to lbry functions
     """
@@ -864,94 +866,8 @@ class Daemon(Daemon_base, Daemon_get, Daemon_settings, Daemon_wallet,
     TRANSACTION_DOC = """
     Transaction management.
     """
-
-    @requires(WALLET_COMPONENT)
-    def jsonrpc_transaction_list(self, account_id=None, wallet_id=None, page=None, page_size=None):
-        """
-        List transactions belonging to wallet
-
-        Usage:
-            transaction_list [<account_id> | --account_id=<account_id>] [--wallet_id=<wallet_id>]
-                             [--page=<page>] [--page_size=<page_size>]
-
-        Options:
-            --account_id=<account_id>  : (str) id of the account to query
-            --wallet_id=<wallet_id>    : (str) restrict results to specific wallet
-            --page=<page>              : (int) page to return during paginating
-            --page_size=<page_size>    : (int) number of items on page during pagination
-
-        Returns:
-            (list) List of transactions
-
-            {
-                "claim_info": (list) claim info if in txn [{
-                                                        "address": (str) address of claim,
-                                                        "balance_delta": (float) bid amount,
-                                                        "amount": (float) claim amount,
-                                                        "claim_id": (str) claim id,
-                                                        "claim_name": (str) claim name,
-                                                        "nout": (int) nout
-                                                        }],
-                "abandon_info": (list) abandon info if in txn [{
-                                                        "address": (str) address of abandoned claim,
-                                                        "balance_delta": (float) returned amount,
-                                                        "amount": (float) claim amount,
-                                                        "claim_id": (str) claim id,
-                                                        "claim_name": (str) claim name,
-                                                        "nout": (int) nout
-                                                        }],
-                "confirmations": (int) number of confirmations for the txn,
-                "date": (str) date and time of txn,
-                "fee": (float) txn fee,
-                "support_info": (list) support info if in txn [{
-                                                        "address": (str) address of support,
-                                                        "balance_delta": (float) support amount,
-                                                        "amount": (float) support amount,
-                                                        "claim_id": (str) claim id,
-                                                        "claim_name": (str) claim name,
-                                                        "is_tip": (bool),
-                                                        "nout": (int) nout
-                                                        }],
-                "timestamp": (int) timestamp,
-                "txid": (str) txn id,
-                "update_info": (list) update info if in txn [{
-                                                        "address": (str) address of claim,
-                                                        "balance_delta": (float) credited/debited
-                                                        "amount": (float) absolute amount,
-                                                        "claim_id": (str) claim id,
-                                                        "claim_name": (str) claim name,
-                                                        "nout": (int) nout
-                                                        }],
-                "value": (float) value of txn
-            }
-
-        """
-        wallet = self.wallet_manager.get_wallet_or_default(wallet_id)
-        if account_id:
-            account = wallet.get_account_or_error(account_id)
-            transactions = account.get_transaction_history
-            transaction_count = account.get_transaction_history_count
-        else:
-            transactions = partial(
-                self.ledger.get_transaction_history, wallet=wallet, accounts=wallet.accounts)
-            transaction_count = partial(
-                self.ledger.get_transaction_history_count, wallet=wallet, accounts=wallet.accounts)
-        return paginate_rows(transactions, transaction_count, page, page_size, read_only=True)
-
-    @requires(WALLET_COMPONENT)
-    def jsonrpc_transaction_show(self, txid):
-        """
-        Get a decoded transaction from a txid
-
-        Usage:
-            transaction_show (<txid> | --txid=<txid>)
-
-        Options:
-            --txid=<txid>  : (str) txid of the transaction
-
-        Returns: {Transaction}
-        """
-        return self.wallet_manager.get_transaction(txid)
+    # jsonrpc_transaction_list
+    # jsonrpc_transaction_show
 
     TXO_DOC = """
     List and sum transaction outputs.
